# app/Modules/Core/OpenApi/pharmanp.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PharmaNP API",
    "version": "1.0.4",
    "description": "PharmaNP ERP API. React, Swagger, and mobile clients use signed JWT bearer tokens."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "PharmaNP API"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate and issue a JWT bearer token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" },
                  "device_name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JwtLoginResponse" },
                "examples": {
                  "success": {
                    "summary": "JWT bearer token response",
                    "value": {
                      "status": "success",
                      "code": 200,
                      "message": "Authenticated.",
                      "data": null,
                      "token_type": "Bearer",
                      "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...",
                      "expires_at": "2026-05-05T17:30:00+05:45"
                    }
                  }
                }
              }
            }
          },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Revoke the current JWT bearer token",
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseEnvelope" },
                "examples": {
                  "success": {
                    "value": {
                      "status": "success",
                      "code": 200,
                      "message": "Logged out.",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthenticated": {
        "description": "Authentication is required."
      },
...
      "ValidationError": {
        "description": "Validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "ApiResponseEnvelope": {
        "type": "object",
        "required": ["status", "code", "message", "data"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "error"],
            "example": "success"
          },
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Records retrieved successfully."
          },
          "data": {
            "nullable": true,
            "description": "Resource object, collection, or null for command-only responses."
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "example": {
          "status": "success",
          "code": 200,
          "message": "Products retrieved successfully.",
          "data": [
            {
              "id": 1,
              "name": "Paracetamol 500",
              "product_code": "ITEM-00001",
              "hs_code": "3004.90",
              "stock_on_hand": 120
            }
          ],
          "links": {
            "first": "https://pharmanp.example.test/api/v1/inventory/products?page=1",
            "last": "https://pharmanp.example.test/api/v1/inventory/products?page=8",
            "prev": null,
            "next": "https://pharmanp.example.test/api/v1/inventory/products?page=2"
          },
          "meta": {
            "current_page": 1,
            "from": 1,
            "last_page": 8,
            "path": "https://pharmanp.example.test/api/v1/inventory/products",
            "per_page": 15,
            "to": 15,
            "total": 120
          }
        }
      },
      "PaginationLinks": {
        "type": "object",
        "nullable": true,
        "properties": {
          "first": { "type": "string", "nullable": true },
          "last": { "type": "string", "nullable": true },
          "prev": { "type": "string", "nullable": true },
          "next": { "type": "string", "nullable": true }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "nullable": true,
        "properties": {
          "current_page": { "type": "integer" },
          "from": { "type": "integer", "nullable": true },
          "last_page": { "type": "integer" },
          "path": { "type": "string" },
          "per_page": { "type": "integer", "example": 15 },
          "to": { "type": "integer", "nullable": true },
          "total": { "type": "integer" }
        }
      },
      "JwtLoginResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponseEnvelope" },
          {
            "type": "object",
            "required": ["token_type", "token", "expires_at"],
            "properties": {
              "token_type": { "type": "string", "example": "Bearer" },
              "token": {
                "type": "string",
                "description": "Signed JWT bearer token used by React, Swagger, and mobile clients."
              },
              "expires_at": { "type": "string", "format": "date-time" }
            }
          }
        ]
      },
...
    }
  },
...
}